refactor: 124-refactor code
