+++ src/helpers/buttons-emitter.ts
@@ -1,12 +1,8 @@
-import { getUserInfo } from '../app/api/get-user-info';
-import { getAuthToken } from '../app/ecommerce/get-auth-token';
-import { CustomEventEmitter, CustomEventEmitterAsync } from '../utils/event-emitter';
-import { updatePersonalData } from '../app/api';
-import type { FetchError } from '../app/types';
+import { CustomEventEmitter } from '../utils/event-emitter';
 import type { WrappedInput } from '../shared/components/input';
+import { resetInputDisplayFromServer } from './reset-input-display-from-server';
 
 export const buttonEmitter = new CustomEventEmitter();
-export const personalDataEmitterAsync = new CustomEventEmitterAsync();
 
 function setButtonState(button: HTMLButtonElement, isActive: boolean): void {
   button.classList.toggle('cursor-pointer', isActive);
@@ -20,31 +16,29 @@ function setInputState(input: HTMLInputElement, isActive: boolean): void {
   input.disabled = !isActive;
 }
 
-async function resetInputDisplayFromServer(
-  firstNameInput: HTMLInputElement,
-  lastNameInput: HTMLInputElement,
-  dateOfBirthInput: HTMLInputElement,
-  emailInput: HTMLInputElement,
-): Promise<FetchError | void> {
-  //! Delete in the future when I save token in local/session storage
-  const token = await getAuthToken('ivanIvanov@yandex.ru', 'Ivan12345');
-
-  if (typeof token !== 'string') {
-    return { message: 'Failed to get token' };
-  }
-
-  const user = await getUserInfo(token);
-  if (!('id' in user)) {
-    return { message: 'Failed to get Personal Info' };
-  }
-
-  firstNameInput.value = user.firstName || '';
-  lastNameInput.value = user.lastName || '';
-  dateOfBirthInput.value = user.dateOfBirth || '';
-  emailInput.value = user.email || '';
+function toggleButtons(
+  edit: HTMLButtonElement,
+  save: HTMLButtonElement,
+  cancel: HTMLButtonElement,
+  isEditMode: boolean,
+): void {
+  setButtonState(edit, !isEditMode);
+  setButtonState(save, isEditMode);
+  setButtonState(cancel, isEditMode);
 }
 
-// eslint-disable-next-line max-lines-per-function
+function toggleInputs(inputs: HTMLInputElement[], isActive: boolean): void {
+  for (const input of inputs) {
+    setInputState(input, isActive);
+  }
+}
+
+function clearErrors(wrappers: WrappedInput[]): void {
+  for (const wrapper of wrappers) {
+    wrapper.errorContainer.textContent = '';
+  }
+}
+
 export function activateButtonEmitter(
   editButton: HTMLButtonElement,
   saveButton: HTMLButtonElement,
@@ -54,91 +48,29 @@ export function activateButtonEmitter(
   dateOfBirthWrapper: WrappedInput,
   emailWrapper: WrappedInput,
 ): void {
-  const firstNameInput = firstNameWrapper.input;
-  const lastNameInput = lastNameWrapper.input;
-  const dateOfBirthInput = dateOfBirthWrapper.input;
-  const emailInput = emailWrapper.input;
+  const wrappers = [firstNameWrapper, lastNameWrapper, dateOfBirthWrapper, emailWrapper];
+  const inputs = [firstNameWrapper.input, lastNameWrapper.input, dateOfBirthWrapper.input, emailWrapper.input];
 
   buttonEmitter.subscribe('editBtnClick', () => {
-    setButtonState(editButton, false);
-    setButtonState(saveButton, true);
-    setButtonState(cancelButton, true);
-
-    setInputState(firstNameInput, true);
-    setInputState(lastNameInput, true);
-    setInputState(dateOfBirthInput, true);
-    setInputState(emailInput, true);
+    toggleButtons(editButton, saveButton, cancelButton, true);
+    toggleInputs(inputs, true);
   });
 
   buttonEmitter.subscribe('saveBtnClick', () => {
-    setButtonState(editButton, true);
-    setButtonState(saveButton, false);
-    setButtonState(cancelButton, false);
-
-    setInputState(firstNameInput, false);
-    setInputState(lastNameInput, false);
-    setInputState(dateOfBirthInput, false);
-    setInputState(emailInput, false);
+    toggleButtons(editButton, saveButton, cancelButton, false);
+    toggleInputs(inputs, false);
   });
 
   buttonEmitter.subscribe('cancelBtnClick', async () => {
-    setButtonState(editButton, true);
-    setButtonState(saveButton, false);
-    setButtonState(cancelButton, false);
+    toggleButtons(editButton, saveButton, cancelButton, false);
+    toggleInputs(inputs, false);
+    clearErrors(wrappers);
 
-    setInputState(firstNameInput, false);
-    setInputState(lastNameInput, false);
-    setInputState(dateOfBirthInput, false);
-    setInputState(emailInput, false);
-
-    firstNameWrapper.errorContainer.textContent = '';
-    lastNameWrapper.errorContainer.textContent = '';
-    dateOfBirthWrapper.errorContainer.textContent = '';
-    emailWrapper.errorContainer.textContent = '';
-
-    await resetInputDisplayFromServer(firstNameInput, lastNameInput, dateOfBirthInput, emailInput);
-  });
-}
-
-export async function updatePersonalDataEmitter(
-  nameInput: HTMLInputElement,
-  surnameInput: HTMLInputElement,
-  birthdateInput: HTMLInputElement,
-  emailInput: HTMLInputElement,
-): Promise<FetchError | void> {
-  personalDataEmitterAsync.subscribe('updateUserData', async () => {
-    const name = nameInput.value;
-    const surname = surnameInput.value;
-    const dateOfBirth = birthdateInput.value;
-    const email = emailInput.value;
-
-    try {
-      //! Delete in the future when I save token via local/session storage
-      const token = await getAuthToken('ivanIvanov@yandex.ru', 'Ivan12345');
-
-      if (typeof token !== 'string') {
-        return { message: 'Failed to get token to update Personal Data' };
-      }
-
-      const user = await getUserInfo(token);
-
-      if (!('id' in user)) {
-        return { message: 'Failed to get User Data' };
-      }
-
-      const updatedUser = await updatePersonalData(name, surname, dateOfBirth, email, user.version, token);
-
-      if (!('id' in updatedUser)) {
-        throw new Error(updatedUser.message);
-      }
-
-      console.log('Updated user', updatedUser);
-    } catch (error) {
-      if (error instanceof Error) {
-        throw new TypeError(error.message);
-      }
-
-      throw new Error('Unexpected error');
-    }
+    await resetInputDisplayFromServer(
+      firstNameWrapper.input,
+      lastNameWrapper.input,
+      dateOfBirthWrapper.input,
+      emailWrapper.input,
+    );
   });
 }
